Remove long double from Log functions and array functions

+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/system/LogInt.java
@@ -64,7 +64,7 @@ public class LogInt extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         //here we cast second parameter to int as anyway it only has few log levels
-        LogUtil.log(logger, (int) ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
+        LogUtil.log(logger, ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
         return VOID_RETURN;
     }
 }
